# backend/Sistema_Crud_API_PIBIC/src/main/java/com/SistemaApiCrud/SistemaCrud/dto/AlternativaPerguntaDTO.java
package com.SistemaApiCrud.SistemaCrud.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class AlternativaPerguntaDTO {

    private Long id;

    @NotBlank(message = "A letra da alternativa e obrigatoria")
    @Size(max = 10, message = "A letra deve ter no maximo 10 caracteres")
    private String letra;

    @NotBlank(message = "O texto da alternativa e obrigatorio")
    @Size(max = 10000, message = "O texto da alternativa deve ter no maximo 10000 caracteres")
    private String texto;

    private Boolean correta;
}