# backend/Sistema_Crud_API_PIBIC/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: SistemaAPI PIBIC
  version: 1.0.0
  description: |
    API REST para gerenciamento e resolucao de casos clinicos academicos.
    As operacoes de IA aceitam somente dados sinteticos ou previamente
    desidentificados e toda saida exige revisao humana antes da publicacao.
    O recurso e educacional e nao substitui decisao clinica.
  contact:
    name: Equipe SistemaAPI PIBIC
  license:
    name: Uso academico
...
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /pacientes:
    get:
      summary: Lista pacientes acessiveis ao usuario autenticado
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Pagina de pacientes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaPaciente"
        "403":
          $ref: "#/components/responses/AcessoNegado"
    post:
      summary: Cria paciente em um caso clinico em rascunho
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Paciente"
      responses:
        "201":
          description: Paciente criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paciente"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
  /pacientes/{id}:
    get:
      summary: Busca paciente por ID
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Paciente encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paciente"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    put:
      summary: Atualiza paciente de um caso clinico em rascunho
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Paciente"
      responses:
        "200":
          description: Paciente atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paciente"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
    delete:
      summary: Remove paciente de um caso clinico em rascunho
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          description: Paciente removido
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
  /conteudos:
    get:
      summary: Lista conteudos clinicos acessiveis ao usuario autenticado
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Pagina de conteudos clinicos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaConteudoClinico"
        "403":
          $ref: "#/components/responses/AcessoNegado"
    post:
      summary: Cria conteudo clinico em um caso em rascunho
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConteudoClinico"
      responses:
        "201":
          description: Conteudo clinico criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConteudoClinico"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
  /conteudos/{id}:
    get:
      summary: Busca conteudo clinico por ID
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Conteudo clinico encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConteudoClinico"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    put:
      summary: Atualiza conteudo clinico de um caso em rascunho
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConteudoClinico"
      responses:
        "200":
          description: Conteudo clinico atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConteudoClinico"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
    delete:
      summary: Remove conteudo clinico de um caso em rascunho
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          description: Conteudo clinico removido
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
  /perguntas:
    get:
      summary: Lista perguntas acessiveis ao usuario autenticado
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Pagina de perguntas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaPergunta"
    post:
      summary: Cria pergunta no caso informado no corpo
      description: O campo idCaso e obrigatorio neste endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerguntaRequest"
      responses:
        "201":
          description: Pergunta criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerguntaResponse"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
  /perguntas/{id}:
    get:
      summary: Busca pergunta por ID
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Pergunta encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerguntaResponse"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    put:
      summary: Atualiza pergunta de um caso em rascunho
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerguntaRequest"
      responses:
        "200":
          description: Pergunta atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerguntaResponse"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
    delete:
      summary: Remove pergunta de um caso em rascunho
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "204":
          description: Pergunta removida
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
  /perguntas/caso/{casoId}:
    get:
      summary: Lista perguntas de um caso com paginacao
      parameters:
        - $ref: "#/components/parameters/casoId"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Pagina de perguntas do caso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaPergunta"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    post:
      summary: Cria pergunta no caso indicado no caminho
      description: O caso do caminho e usado; idCaso no corpo e opcional.
      parameters:
        - $ref: "#/components/parameters/casoId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerguntaRequest"
      responses:
        "201":
          description: Pergunta criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerguntaResponse"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    casoId:
      name: casoId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    page:
      name: page
      in: query
      description: Indice da pagina, iniciando em zero.
      schema:
        type: integer
        minimum: 0
        default: 0
    size:
      name: size
      in: query
      description: Quantidade de elementos por pagina.
      schema:
        type: integer
        minimum: 1
        default: 20
    sort:
      name: sort
      in: query
      description: Ordenacao no formato campo,direcao.
      schema:
        type: string
        example: nome,asc
  responses:
...
    CorpoRequisicaoExcedido:
      description: O corpo da requisicao excede o limite configurado.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: urn:sistema-api-pibic:problem:corpo-requisicao-excedido
            title: Corpo da requisicao excedido
            status: 413
            detail: O corpo da requisicao excede o limite permitido
            instance: /auth/login
            timestamp: "2026-08-22T20:00:00Z"
            correlationId: 7c973582-36bc-45a7-980f-285901008ad8
            erro: O corpo da requisicao excede o limite permitido
...
    DadosInvalidos:
      description: A requisicao possui campos, JSON, parametros ou tipos invalidos.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    AcessoNegado:
      description: O usuario autenticado nao possui permissao para a operacao.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NaoEncontrado:
      description: O recurso solicitado nao foi encontrado.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflito:
      description: O estado atual ou a concorrencia dos dados impede a operacao.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
        Retry-After:
          description: Segundos ate a repeticao recomendada quando o conflito for transitorio.
          schema:
            type: integer
            format: int64
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  headers:
    CorrelationId:
      description: Identificador da requisicao para rastreamento ponta a ponta.
      schema:
        type: string
        minLength: 1
        maxLength: 64
  schemas:
...
    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - instance
        - timestamp
        - correlationId
        - erro
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        timestamp:
          type: string
          format: date-time
        correlationId:
          type: string
        erro:
          type: string
          deprecated: true
          description: Alias legado de detail; mantido temporariamente para compatibilidade.
        errors:
          type: object
          additionalProperties:
            type: string
          description: Erros por campo quando a validacao falha.
        campos:
          type: object
          deprecated: true
          additionalProperties:
            type: string
          description: Alias legado de errors; mantido temporariamente para compatibilidade.
...
    Paciente:
      type: object
      required:
        - idCaso
        - nome
        - profissao
        - sexo
        - idade
        - estadoCivil
        - altura
        - peso
      properties:
        idPaciente:
          type: integer
          format: int64
          readOnly: true
        idCaso:
          type: integer
          format: int64
          minimum: 1
        nome:
          type: string
          maxLength: 150
        profissao:
          type: string
          maxLength: 120
        sexo:
          type: string
          enum: [MASCULINO, FEMININO, OUTRO, NAO_INFORMADO]
        idade:
          type: integer
          minimum: 0
          maximum: 130
        estadoCivil:
          type: string
          enum: [SOLTEIRO, CASADO, DIVORCIADO, VIUVO, SEPARADO, UNIAO_ESTAVEL, NAO_INFORMADO]
        altura:
          type: string
          maxLength: 20
        peso:
          type: string
          maxLength: 20
    ConteudoClinico:
      type: object
      required:
        - idCaso
        - sintomas
        - contexto
        - examClinico
        - antecClinico
        - diagEsperado
      properties:
        idConteudo:
          type: integer
          format: int64
          readOnly: true
        idCaso:
          type: integer
          format: int64
          minimum: 1
        sintomas:
          type: string
          maxLength: 10000
        contexto:
          type: string
          maxLength: 10000
        examClinico:
          type: string
          maxLength: 10000
        antecClinico:
          type: string
          maxLength: 10000
        diagEsperado:
          type: string
          maxLength: 10000
...
    AlternativaPergunta:
      type: object
      properties:
        id:
          type: integer
          format: int64
        letra:
          type: string
          maxLength: 10
        texto:
          type: string
          maxLength: 10000
        correta:
          type: boolean
...
    RubricaPergunta:
      type: object
      additionalProperties: false
      description: >-
        Rubrica estruturada opcional para perguntas DISCURSIVA e CONDUTA_CLINICA.
        A propriedade resposta permanece obrigatoria como resumo e compatibilidade legada.
      properties:
        criteriosEssenciais:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 2000
        criteriosPontuacao:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 2000
        errosGraves:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 2000
        justificativas:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 2000
        prioridades:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 2000
        sinaisEscalonamento:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 2000
    PerguntaRequest:
      type: object
      required:
        - texto
        - tipo
        - gabarito
        - resposta
      properties:
        idCaso:
          type: integer
          format: int64
          nullable: true
        texto:
          type: string
          maxLength: 10000
        tipo:
          type: string
          enum: [MULTIPLA_ESCOLHA, DISCURSIVA, VERDADEIRO_FALSO]
          example: MULTIPLA_ESCOLHA
        gabarito:
          type: string
          maxLength: 10000
        resposta:
          type: string
          maxLength: 10000
        rubrica:
          allOf:
            - $ref: "#/components/schemas/RubricaPergunta"
          nullable: true
        alternativas:
          type: array
          maxItems: 5
          items:
            $ref: "#/components/schemas/AlternativaPergunta"
    PerguntaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        idCaso:
          type: integer
          format: int64
        texto:
          type: string
        tipo:
          type: string
          enum: [MULTIPLA_ESCOLHA, DISCURSIVA, VERDADEIRO_FALSO, DIAGNOSTICO, CONDUTA_CLINICA]
          example: MULTIPLA_ESCOLHA
        gabarito:
          type: string
        resposta:
          type: string
        rubrica:
          allOf:
            - $ref: "#/components/schemas/RubricaPergunta"
          nullable: true
        alternativas:
          type: array
          items:
            $ref: "#/components/schemas/AlternativaPergunta"
...
    PaginaBase:
      type: object
      required:
        - totalElements
        - totalPages
        - number
        - size
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
...
    PaginaPaciente:
      allOf:
        - $ref: "#/components/schemas/PaginaBase"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/Paciente"
    PaginaConteudoClinico:
      allOf:
        - $ref: "#/components/schemas/PaginaBase"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/ConteudoClinico"
    PaginaPergunta:
      allOf:
        - $ref: "#/components/schemas/PaginaBase"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/PerguntaResponse"
...